Created a multi-module project in order to add a client application making oauth2 calls
