+++ docker-compose.yaml
@@ -10,30 +10,43 @@ services:
       - "5433:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
-      - ./init-db.sh:/docker-entrypoint-initdb.d/init-db.sh
+      - ./oauth2-resource-server/init-db.sh:/docker-entrypoint-initdb.d/init-db.sh
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U productuser -d productdb"]
+      interval: 10s
+      timeout: 5s
+      retries: 5
 
   keycloak:
     image: quay.io/keycloak/keycloak:20.0.3
     container_name: keycloak
     environment:
-      KEYCLOAK_USER: admin
-      KEYCLOAK_PASSWORD: admin
       KEYCLOAK_ADMIN: admin
       KEYCLOAK_ADMIN_PASSWORD: admin
-      DB_VENDOR: h2
-      KEYCLOAK_HTTP_HOST: 0.0.0.0  # Replaces `-b` option
-      KEYCLOAK_HTTP_PORT: 8080  # Use HTTP port, not HTTPS
-      KEYCLOAK_IMPORT: /opt/keycloak/realms/keycloak-realm.json  # Import your realm file
-      KEYCLOAK_DEV_MODE: "true"  # Enable development mode
+      KC_DB: dev-file
+      KC_HTTP_RELATIVE_PATH: /
+      KC_HOSTNAME_STRICT: "false"
+      KC_HOSTNAME_STRICT_HTTPS: "false"
+      KC_HTTP_ENABLED: "true"
+      KC_HEALTH_ENABLED: "true"
     ports:
       - "8081:8080"
     volumes:
-      - ./keycloak-realm.json:/opt/keycloak/realms/keycloak-realm.json  # Mount your realm JSON file
-    command: ["start-dev"]
+      - ./keycloak-realm.json:/opt/keycloak/data/import/realm.json
+    command:
+      - start-dev
+      - --import-realm
+    healthcheck:
+      test: [ "CMD", "curl", "-f", "http://localhost:8080/health/ready" ]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+      start_period: 30s
 
-  app:
-    build: .
-    container_name: oauth2product
+  resource-server:
+    build:
+      context: ./oauth2-resource-server
+    container_name: oauth2-resource-server
     environment:
       SPRING_DATASOURCE_URL: jdbc:postgresql://postgres:5432/productdb
       SPRING_DATASOURCE_USERNAME: productuser
@@ -43,9 +56,23 @@ services:
     ports:
       - "8080:8080"
     depends_on:
-      - postgres
-      - keycloak
-    entrypoint: ["/bin/bash", "/app/wait-for-it.sh", "postgres:5432", "--", "java", "-jar", "/app/app.jar"]
+      postgres:
+        condition: service_healthy
+      keycloak:
+        condition: service_healthy
+
+  oauth2-client:
+    build:
+      context: ./oauth2-client
+    container_name: oauth2-client
+    environment:
+      SPRING_SECURITY_OAUTH2_CLIENT_PROVIDER_KEYCLOAK_ISSUER_URI: http://keycloak:8080/realms/product-realm
+    ports:
+      - "8082:8082"
+    depends_on:
+      keycloak:
+        condition: service_healthy
 
 volumes:
   postgres_data:
+  keycloak_data:
