+++ oauth2-client/src/main/java/dev/sharanggupta/Main.java
@@ -1,0 +1,7 @@
+package dev.sharanggupta;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
