+++ oauth2-resource-server/src/main/java/dev/sharanggupta/oauth2resourceserver/Oauth2productApplication.java
@@ -1,0 +1,13 @@
+package dev.sharanggupta.oauth2resourceserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Oauth2productApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Oauth2productApplication.class, args);
+	}
+
+}
