+++ oauth2-resource-server/src/main/java/dev/sharanggupta/oauth2resourceserver/exception/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package dev.sharanggupta.oauth2resourceserver.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+}
