+++ oauth2-resource-server/src/main/java/dev/sharanggupta/oauth2resourceserver/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package dev.sharanggupta.oauth2resourceserver.repository;
+
+import dev.sharanggupta.oauth2resourceserver.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {}
