+++ oauth2-resource-server/src/main/java/dev/sharanggupta/oauth2resourceserver/service/ProductService.java
@@ -1,0 +1,42 @@
+package dev.sharanggupta.oauth2resourceserver.service;
+
+import dev.sharanggupta.oauth2resourceserver.domain.Product;
+import dev.sharanggupta.oauth2resourceserver.exception.ProductNotFoundException;
+import dev.sharanggupta.oauth2resourceserver.repository.ProductRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ProductService {
+    private final ProductRepository repository;
+
+    public ProductService(ProductRepository repository) {
+        this.repository = repository;
+    }
+
+    public List<Product> findAll() {
+        return repository.findAll();
+    }
+    public Product findById(Long id) {
+        return repository.findById(id).orElseThrow(() -> new ProductNotFoundException("Product not found"));
+    }
+
+    public Product create(Product product) {
+        return repository.save(product);
+    }
+
+    public Product update(Long id, Product product) {
+        Product existingProduct = repository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException("Product " + id + " not found"));
+        existingProduct.setName(product.getName());
+        existingProduct.setPrice(product.getPrice());
+        return repository.save(existingProduct);
+    }
+
+    public void deleteById(Long id) {
+        Product product = repository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException("Product " + id + " not found"));
+        repository.delete(product);
+    }
+}
