+++ oauth2-resource-server/src/main/resources/application.yaml
