+++ oauth2-resource-server/src/main/resources/db/migration/V1__create_product_table.sql
