+++ oauth2-resource-server/src/test/java/dev/sharanggupta/oauth2resourceserver/Oauth2productApplicationTests.java
@@ -1,0 +1,18 @@
+package dev.sharanggupta.oauth2resourceserver;
+
+import dev.sharanggupta.oauth2resourceserver.config.TestContainersConfig;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ActiveProfiles;
+
+@Import(TestContainersConfig.class)
+@ActiveProfiles("test")
+@SpringBootTest
+class Oauth2productApplicationTests extends BaseIntegrationTest{
+
+	@Test
+	void contextLoads() {
+		// This test is empty because it is only used to check if the Spring context loads successfully
+	}
+}
