+++ oauth2-resource-server/src/test/java/dev/sharanggupta/oauth2resourceserver/config/TestContainersConfig.java
@@ -1,0 +1,7 @@
+package dev.sharanggupta.oauth2resourceserver.config;
+
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+public class TestContainersConfig {
+}
