+++ oauth2-resource-server/src/test/resources/db/migration/V1__create_product_table.sql
