--- src/main/java/dev/sharanggupta/oauth2product/Oauth2productApplication.java
@@ -1,13 +1,0 @@
-package dev.sharanggupta.oauth2product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Oauth2productApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Oauth2productApplication.class, args);
-	}
-
-}
