--- src/main/java/dev/sharanggupta/oauth2product/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package dev.sharanggupta.oauth2product.exception;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
