--- src/test/java/dev/sharanggupta/oauth2product/Oauth2productApplicationTests.java
@@ -1,18 +1,0 @@
-package dev.sharanggupta.oauth2product;
-
-import dev.sharanggupta.oauth2product.config.TestContainersConfig;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ActiveProfiles;
-
-@Import(TestContainersConfig.class)
-@ActiveProfiles("test")
-@SpringBootTest
-class Oauth2productApplicationTests extends BaseIntegrationTest{
-
-	@Test
-	void contextLoads() {
-		// This test is empty because it is only used to check if the Spring context loads successfully
-	}
-}
