--- src/test/java/dev/sharanggupta/oauth2product/config/TestContainersConfig.java
@@ -1,7 +1,0 @@
-package dev.sharanggupta.oauth2product.config;
-
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers
-public class TestContainersConfig {
-}
